requirements.txt
Flask
requests
mysql-connector
plotly
pandas